% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buy_bonds.R
\name{buy_bonds}
\alias{buy_bonds}
\title{Buy Bonds}
\usage{
buy_bonds(investment_vector, withdrawal_vector, coupon, tax_rate)
}
\arguments{
\item{investment_vector}{a vector of annual investments}

\item{withdrawal_vector}{a vector of annual withdrawals}

\item{coupon}{the interest rate}

\item{tax_rate}{the investor's average tax rate charged on interest payments}
}
\value{
data.frame with summary statistics for each window
}
\description{
Shows the portfolio performance of buying one or more bonds
}
\details{
What happens to the portfolio if each investment is spent on buying
a bond that yields a rate of 'coupon' discounted by 'tax_rate'
}
\examples{
\dontrun{
investment_vector <- seq(1,30)*10
withdrawal_vector <- c(investment_vector[1:10]  * 0.15,
                   investment_vector[11:20] * 0.35,
                   investment_vector[21:30] * runif(10, min=0.01, max=0.90))
coupon   <- 0.0261 # 30-year rate in May 2016
tax_rate <- 0.14   # Mitt Romney's average rate
bond_df <- buy_bonds <- function(investment_vector,
                                 withdrawal_vector,
                                 coupon,
                                 tax_rate)
}

}
\author{
George Fisher
}

