% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_chg.R
\name{calc_chg}
\alias{calc_chg}
\title{Calculate The Changes In A Series}
\usage{
calc_chg(series)
}
\arguments{
\item{series}{a vector of values, assumed to be an evenly-spaced time series,
to calculate the changes between}
}
\value{
a vector of the changes with an NA in the first position
}
\description{
Calculate The Changes In A Series
}
\details{
The amount of change between two numbers in a series is current / previous.
}
\examples{
series = c(16.66, 16.85, 16.93, 16.98, 17.08, 17.03)
calc_chg(series)

}
\author{
George Fisher

#@description
}

