% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt}
\alias{fmt}
\title{Format a Number with Commas and No Decimals}
\usage{
fmt(dat)
}
\arguments{
\item{dat}{number to be formatted}
}
\value{
formatted string
}
\description{
Format a Number with Commas and No Decimals
}
\examples{
\dontrun{
dat = 123456.654
fmt(dat)
}

}
\author{
George Fisher

#@description

#@details
}

