% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effect_of_fees.R
\name{plot_effect_of_fees}
\alias{plot_effect_of_fees}
\title{Plot the Effect of Fees}
\usage{
plot_effect_of_fees(raw_data, annual_fee, output_path)
}
\arguments{
\item{raw_data}{s&p500 daily data}

\item{annual_fee}{annual fee}

\item{output_path}{file path of plot}
}
\value{
No explicit returned values, just a plot to output_path
}
\description{
Plot the Effect of Fees
}
\author{
George Fisher

#@description

#@details

#@examples
}

