% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ending_bal_distribution.R
\name{plot_ending_bal_distribution}
\alias{plot_ending_bal_distribution}
\title{Plot the Distribution of Ending Balances}
\usage{
plot_ending_bal_distribution(window_df, output_path, window_width)
}
\arguments{
\item{window_df}{data.frame output of sp500SlidingWindow}

\item{output_path}{folder to write the file and the plot, if NULL
then the image is plotted to the screen}

\item{window_width}{number of years in each window}
}
\value{
No explicit returned values, (1) writes 'distribution.txt' with
statistics about the windows (2) creates a plot 'distribution.png' in the
output_path showing the distribution of outcomes
}
\description{
Plot the Distribution of Ending Balances
}
\author{
George Fisher

#@description

#@details

#@examples
}

