% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ending_bal_withdrawals.R
\name{plot_ending_bal_withdrawals}
\alias{plot_ending_bal_withdrawals}
\title{Plot the Ending Balances and Withdrawals}
\usage{
plot_ending_bal_withdrawals(window_df, output_path, lump_sum, annual_withdrawal,
  annual_inflation, window_width)
}
\arguments{
\item{window_df}{data.frame output of sp500SlidingWindow}

\item{output_path}{folder to write the file and the plot}

\item{lump_sum}{the amount initiall invested}

\item{annual_withdrawal}{the amount of hoped-for withdrawals}

\item{annual_inflation}{each year the withdrawal goes up this much}

\item{window_width}{the number of years in the sliding window}
}
\value{
No explicit returned values, creates the plots
'ending_bal.png' and 'withdrawals.png' in the output_path
}
\description{
Plot the Ending Balances and Withdrawals
}
\author{
George Fisher

#@description

#@details

#@examples
}

