% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_market_results.R
\name{plot_market_results}
\alias{plot_market_results}
\title{Plot The Daily Chart For One Window}
\usage{
plot_market_results(filtered_data, statistics, output_path)
}
\arguments{
\item{filtered_data}{data.frame of the window's daily data}

\item{statistics}{summary statistics for this window}

\item{output_path}{file path to write the graph to}
}
\value{
No explicit returned values, merely the side-effect of a graphs
sent to the output_path folder
}
\description{
Plot The Daily Chart For One Window
}
\author{
George Fisher

#@description

#@details

#@examples
}

