% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_total_stock_market.R
\name{plot_total_stock_market}
\alias{plot_total_stock_market}
\title{Plot the Entire Stock Market History from 1950}
\usage{
plot_total_stock_market(raw_data, output_path)
}
\arguments{
\item{raw_data}{output of SP500TR_1950()}

\item{output_path}{folder to write the file and the plot}
}
\value{
No explicit returned values, creates a plot
paste0("SP500_", start_year, "-", end_year, ".png") in the
output_path
}
\description{
Plot the Entire Stock Market History from 1950
}
\author{
George Fisher

#@description

#@details

#@examples
}

